Move HostPacket and NetworkPacket into `Packet.h`
