+++ Sming/Core/Data/Packet.h
@@ -76,3 +76,71 @@ struct Packet {
 		pos += len;
 	}
 };
+
+/**
+ * @brief Helper class for reading/writing packet content in network byte-order (MSB first)
+ */
+struct NetworkPacket : public Packet {
+	using Packet::Packet;
+
+	uint16_t peek16() const
+	{
+		return (data[pos] << 8) | data[pos + 1];
+	}
+
+	uint16_t read16() const
+	{
+		return (read8() << 8) | read8();
+	}
+
+	uint32_t read32() const
+	{
+		return (read16() << 16) | read16();
+	}
+
+	void write16(uint16_t value)
+	{
+		write8(value >> 8);
+		write8(value & 0xff);
+	}
+
+	void write32(uint32_t value)
+	{
+		write16(value >> 16);
+		write16(value & 0xffff);
+	}
+};
+
+/**
+ * @brief Helper class for reading/writing packet content in host byte-order (LSB first)
+ */
+struct HostPacket : public Packet {
+	using Packet::Packet;
+
+	uint16_t peek16() const
+	{
+		return data[pos] | (data[pos + 1] << 8);
+	}
+
+	uint16_t read16() const
+	{
+		return read8() | (read8() << 8);
+	}
+
+	uint32_t read32() const
+	{
+		return read16() | (read16() << 16);
+	}
+
+	void write16(uint16_t value)
+	{
+		write8(value & 0xff);
+		write8(value >> 8);
+	}
+
+	void write32(uint32_t value)
+	{
+		write16(value & 0xffff);
+		write16(value >> 16);
+	}
+};
